backend/requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
email-validator==2.2.0

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
